# conecte-api/src/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Conecte API",
      "version": "1.0.0",
      "description": "API de agendamentos médicos para médicos e pacientes com suporte a WebSocket."
    },
    "servers": [
      { "url": "http://localhost:3000" }
    ],
    "paths": {
      "/auth/register": {
        "post": {
          "summary": "Registrar novo usuário",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string" },
                    "email": { "type": "string", "format": "email" },
                    "password": { "type": "string", "minLength": 6 },
                    "role": { "type": "string", "enum": ["PATIENT", "DOCTOR"] }
                  },
                  "required": ["name", "email", "password", "role"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Usuário registrado com sucesso"
            }
          }
        }
      },
      "/auth/login": {
        "post": {
          "summary": "Autenticar usuário",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string" },
                    "password": { "type": "string" }
                  },
                  "required": ["email", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Login bem-sucedido"
            }
          }
        }
      },
      "/medicos": {
        "get": {
          "summary": "Listar todos os médicos",
          "parameters": [
            {
              "name": "nome",
              "in": "query",
              "schema": { "type": "string" }
            },
            {
              "name": "order",
              "in": "query",
              "schema": { "type": "string", "enum": ["asc", "desc"] }
            }
          ],
          "responses": {
            "200": {
              "description": "Lista de médicos"
            }
          }
        }
      },
      "/medicos/{id}/horarios": {
        "post": {
          "summary": "Criar horário disponível",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "string" }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "dateTime": { "type": "string", "format": "date-time" }
                  },
                  "required": ["dateTime"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Horário criado com sucesso"
            }
          }
        },
        "get": {
          "summary": "Listar horários disponíveis de um médico",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "string" }
            }
          ],
          "responses": {
            "200": {
              "description": "Lista de horários"
            }
          }
        }
      },
      "/agendamentos": {
        "get": {
          "summary": "Listar agendamentos do usuário logado",
          "responses": {
            "200": {
              "description": "Lista de agendamentos com dados do médico e do paciente"
            }
          }
        },
        "post": {
          "summary": "Criar novo agendamento",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "slotId": { "type": "string" }
                  },
                  "required": ["slotId"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Agendamento criado com sucesso"
            }
          }
        }
      }
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [{ "bearerAuth": [] }]
  }
  